Refactor : Draggable 컴포넌트 이동
- core directory로 옮기는 것이 파일의 의미상 더 적합한 것 같아서 경로를 이동했습니다.
- 더 나은 가독성을 위해 css지정부분을 template literal로 수정했습니다.
